requirements.txt
python-gitlab==1.6.0